+++ pynav/lib/models.py
@@ -6,12 +6,10 @@ from pynav.types import (
 )
 from pynav.lib.states import (
     CompositeState,
-    IMUState,
     MatrixLieGroupState,
-    SE23State,
     VectorState,
 )
-from pylie import SO2, SO3, SE23
+from pylie import SO2, SO3
 import numpy as np
 from typing import List
 from scipy.linalg import block_diag
@@ -142,218 +140,6 @@ class BodyFrameVelocity(ProcessModel):
             L = dt * Ad @ x.group.left_jacobian(-u.value * dt)
 
         return L @ self._Q @ L.T
-
-
-class Imu:
-    """
-    Data container for an IMU reading.
-    """
-
-    def __init__(
-        self,
-        gyro: np.ndarray,
-        accel: np.ndarray,
-        stamp: float,
-        bias_gyro_walk=[0, 0, 0],
-        bias_accel_walk=[0, 0, 0],
-    ):
-        self.gyro = np.array(gyro).ravel()
-        self.accel = np.array(accel).ravel()
-        self.bias_gyro_walk = np.array(bias_gyro_walk).ravel()
-        self.bias_accel_walk = np.array(bias_accel_walk).ravel()
-        self.stamp = stamp
-
-    def plus(self, w: np.ndarray):
-        w = w.ravel()
-        self.gyro += w[0:3]
-        self.accel += w[3:6]
-        self.bias_gyro_walk += w[6:9]
-        self.bias_accel_walk += w[9:12]
-
-    def copy(self):
-        return Imu(
-            self.gyro.copy(),
-            self.accel.copy(),
-            self.stamp,
-            self.bias_gyro_walk.copy(),
-            self.bias_accel_walk.copy(),
-        )
-
-
-class IMUKinematics(ProcessModel):
-    def __init__(self, Q: np.ndarray, g_a=[0, 0, -9.80665]):
-        """Instantiate the IMUKinematic model.
-
-        Parameters
-        ----------
-        Q : np.ndarray
-            Discrete-time noise matrix.
-        g_a : np.ndarray
-            Gravity vector resolved in the inertial frame.
-            If None, default value is set to [0; 0; -9.81].
-        """
-        self._Q = Q
-
-        self._gravity = np.array(g_a).reshape((-1, 1))
-
-    def evaluate(self, x: IMUState, u: Imu, dt: float) -> IMUState:
-        """Propagates an IMU state forward one timestep from an IMU measurement.
-
-        The continuous-time IMU equations are discretized using the assumption
-        that the acceleration in the inertial frame is constant between two timesteps.
-
-        TODO: Implement discritization of IMU equations assuming that the measurements
-        are constant over the discritizaiton interval. This is exact if the measurements
-        are actually constant over the interval.
-
-        Parameters
-        ----------
-        x : IMUState
-            Current IMU state
-        u : Imu
-            IMU measurement,
-        dt : float
-            timestep.
-
-        Returns
-        -------
-        IMUState
-            Propagated IMUState.
-        """
-
-        # Get unbiased inputs
-        bias_gyro = x.bias_gyro.reshape((-1, 1))
-        bias_accel = x.bias_accel.reshape((-1, 1))
-
-        unbiased_gyro = u.gyro.reshape((-1, 1)) - bias_gyro
-        unbiased_accel = u.accel.reshape((-1, 1)) - bias_accel
-
-        # Extract states at time km1
-        C_km1 = x.attitude
-        v_km1 = x.velocity.reshape((-1, 1))
-        r_km1 = x.position.reshape((-1, 1))
-
-        g_a = self._gravity
-
-        # Propagate the navigation state forward in time
-        C_ab_k = C_km1 @ SO3.Exp(dt * (unbiased_gyro))
-        r_zw_a_k = (
-            r_km1 + dt * v_km1 + 0.5 * dt * dt * (C_km1 @ unbiased_accel + g_a)
-        )
-        v_zw_a_k = v_km1 + dt * (C_km1 @ unbiased_accel + g_a)
-
-        # Propagate the biases forward in time using random walk
-        if hasattr(u, "bias_gyro_walk"):
-            x.bias_gyro = bias_gyro.ravel() + dt * u.bias_gyro_walk.ravel()
-
-        if hasattr(u, "bias_accel_walk"):
-            x.bias_accel = bias_accel.ravel() + dt * u.bias_accel_walk.ravel()
-
-        x.attitude = C_ab_k
-        x.velocity = v_zw_a_k
-        x.position = r_zw_a_k
-
-        return x
-
-    def jacobian(self, x: IMUState, u: Imu, dt: float) -> np.ndarray:
-        """Computes the discrete-time Jacobian.
-
-        Here, we start in continuous-time and discritize the linearized
-        continuous-time system.
-
-        Parameters
-        ----------
-        x : IMUState
-            The state at which to evaluate the Jacobian at.
-        u : StampedValue
-            Measurement at which to evaluate the Jacobian at.
-        dt : float
-            Discritization timestep.
-
-        Returns
-        -------
-        np.ndarray
-            Process model Jacobian
-        """
-        # Compute unbiased gyro and accel measurements using bias estimates
-
-        # Get unbiased inputs
-        bias_gyro = x.bias_gyro.reshape((-1, 1))
-        bias_accel = x.bias_accel.reshape((-1, 1))
-
-        unbiased_gyro = u.gyro.reshape((-1, 1)) - bias_gyro
-        unbiased_accel = u.accel.reshape((-1, 1)) - bias_accel
-
-        # Continuous-time Jacobian
-        A_ct = np.zeros((x.dof, x.dof))
-
-        # Extract relevant states
-        C = x.attitude
-        v = x.velocity
-        r = x.position
-
-        if x.direction == "left":
-            # TODO: These are the Jacobians for the right-invariant error,
-            # check that they are correct for a left perturbation.
-            A_ct[0:3, 9:12] = -C
-            A_ct[3:6, 0:3] = SO3.cross(self._gravity)
-            A_ct[3:6, 9:12] = -SO3.cross(v) @ C
-            A_ct[3:6, 12:15] = -C
-            A_ct[6:9, 3:6] = np.identity(3)
-            A_ct[6:9, 9:12] = -SO3.cross(r) @ C
-        else:
-            raise NotImplementedError("Right jacobian not implemented.")
-
-        # Approximate matrix exponential to third-order
-        Adt = A_ct * dt
-        Adt_square = Adt @ Adt
-        Adt_cube = Adt_square @ Adt
-        Phi = np.identity(x.dof) + Adt + Adt_square / 2.0 + Adt_cube / 6.0
-
-        return Phi
-
-    def covariance(self, x: IMUState, u: StampedValue, dt: float) -> np.ndarray:
-        """Computes the discrete-time covariance.
-
-        Here again, we start in continuous-time and then discritize the
-        linear system.
-
-        Parameters
-        ----------
-        x : IMUState
-            _description_
-        u : StampedValue
-            _description_
-        dt : float
-            _description_
-
-        Returns
-        -------
-        np.ndarray
-            _description_
-        """
-
-        # Continuous-time L matrix
-        L_ct = np.zeros((x.dof, 12))
-
-        # Extract relevant states
-        C = x.attitude
-        v = x.velocity.reshape((-1, 1))
-        r = x.position.reshape((-1, 1))
-
-        if x.direction == "left":
-            L_ct[0:3, 0:3] = C
-            L_ct[3:6, 0:3] = SO3.cross(v) @ C
-            L_ct[3:6, 3:6] = C
-            L_ct[6:9, 0:3] = SO3.cross(r) @ C
-            L_ct[9:12, 6:9] = -np.identity(3)
-            L_ct[12:15, 9:12] = -np.identity(3)
-        else:
-            raise NotImplementedError("Right jacobian not implemented.")
-
-        # Compute discrete-time noise
-        return L_ct @ self._Q @ L_ct.T * dt**2
-
 
 class RelativeBodyFrameVelocity(ProcessModel):
     def __init__(self, Q1: np.ndarray, Q2: np.ndarray):
@@ -956,218 +742,4 @@ class InvariantMeasurement(Measurement):
             stamp=meas.stamp,
             model=model,
         )
-        return V
 
-    def _U_matrix(self, u:Imu, dt: float):
-        phi = u.gyro * dt
-        O = SO3.Exp(phi)
-        J = SO3.left_jacobian(phi)
-        a = u.accel.reshape((-1,1))
-        V = self._V_matrix(phi)
-        U = np.identity(5)
-        U[:3,:3] = O
-        U[:3,3] = np.ravel(dt* J @ a)
-        U[:3,4] = np.ravel(dt**2 * V @ a)
-        U[3,4] = dt
-        return U
-
-    def _U_matrix_inv(self, u:Imu, dt: float):
-        phi = u.gyro * dt
-        O = SO3.Exp(phi)
-        V = self._V_matrix(phi)
-        J = SO3.left_jacobian(phi)
-        a = u.accel.reshape((-1,1))
-        U_inv = np.identity(5)
-        U_inv[:3,:3] = O.T
-        U_inv[:3,3] = np.ravel(-dt* O.T @ J @ a)
-        U_inv[:3,4] = np.ravel(dt**2 * O.T @ (J - V) @ a)
-        U_inv[3,4] = -dt
-        return U_inv
-
-    def _G_matrix(self, dt):
-        G = np.identity(5)
-        G[:3, 3] = dt * self._g
-        G[:3, 4] = -0.5 * dt**2 * self._g
-        G[3, 4] = -dt
-        return G
-
-    def _G_matrix_inv(self, dt):
-        G_inv = np.identity(5)
-        G_inv[:3, 3] = -dt * self._g
-        G_inv[:3, 4] = - 0.5 * dt**2 * self._g
-        G_inv[3, 4] = dt
-        return G_inv
-
-    # @staticmethod
-    # def _adjoint_U(U):
-    #     O = U[:3,:3]
-    #     dt = U[3,4]
-    #     Ja = U[:3,3].reshape((-1,1))/dt
-    #     Va = U[:3,4].reshape((-1,1))/dt**2
-    #     Ad = np.zeros((9,9))
-    #     Ad[:3,:3] = O
-    #     Ad[3:6,:3] = dt*SO3.wedge(Ja) @ O
-    #     Ad[3:6,3:6] = O
-
-    #     Ad[6:9,:3] = -dt**2*SO3.wedge(Ja - Va) @ O
-    #     Ad[6:9,3:6] = -dt * O 
-    #     Ad[6:9,6:9] = O
-    #     return Ad
-
-    @staticmethod
-    def _adjoint_U(U):
-        O = U[:3,:3]
-        c = U[3,4]
-        a = U[:3,3].reshape((-1,1))
-        b = U[:3,4].reshape((-1,1))
-        Ad = np.zeros((9,9))
-        Ad[:3,:3] = O
-        Ad[3:6,:3] = SO3.wedge(a) @ O
-        Ad[3:6,3:6] = O
-
-class Imu(StampedValue):
-    """
-    A data container for IMU measurements.
-    """
-
-    def __init__(self, accel: np.ndarray, gyro: np.ndarray, stamp: float):
-        """
-
-        Parameters
-        ----------
-        accel : np.ndarray with size 3
-            Accelerometer reading in body frame.
-        gyro : np.ndarray with size 3
-            Gyroscope reading in body frame.
-        stamp : float
-            Timestamp of the measurement.
-        """
-        gyro = np.array(gyro).ravel()
-        accel = np.array(accel).ravel()
-        self.value = [gyro, accel]
-        self.stamp = stamp
-
-    def plus(self, w: np.ndarray):
-        w = np.array(w).ravel()
-        self.value[0] += w[:3]
-        self.value[1] += w[3:]
-
-    @property
-    def gyro(self):
-        return self.value[0]
-
-    @property
-    def accel(self):
-        return self.value[1]
-
-
-class ImuKinematics(ProcessModel):
-    def __init__(self) -> None:
-        self._g = np.array([0, 0, -9.80665])
-
-    def evaluate(self, x: SE23State, u: Imu, dt: float) -> SE23State:
-        G = self._G_matrix(dt)
-
-        U = self._U_matrix(u, dt)
-        x.value = G @ x.value @ U
-        x.stamp += dt
-        return x
-
-    def jacobian(self, x: SE23State, u: Imu, dt: float) -> np.ndarray:
-        if x.direction == "left":
-            G = self._G_matrix(dt)
-            return self._adjoint_U(G)
-        elif x.direction =="right":
-            U_inv = self._U_matrix_inv(u, dt)
-            return self._adjoint_U(U_inv)
-        else:
-            raise ValueError("Invalid direction")
-
-
-    def covariance(self, x: SE23State, u: StampedValue, dt: float) -> np.ndarray:
-        raise NotImplementedError
-
-    @staticmethod
-    def _V_matrix(phi_vec: np.ndarray):
-        phi = np.linalg.norm(phi_vec)
-        a = phi_vec / phi
-        a = a.reshape((-1, 1))
-        a_wedge = SO3.wedge(a)
-        V = (
-            0.5 * a @ a.T
-            + ((1 - np.sin(phi)/phi)) / phi * a_wedge
-            + (np.cos(phi) - 1) / phi**2 * a_wedge @ a_wedge
-        )
-        return V
-
-    def _U_matrix(self, u:Imu, dt: float):
-        phi = u.gyro * dt
-        O = SO3.Exp(phi)
-        J = SO3.left_jacobian(phi)
-        a = u.accel.reshape((-1,1))
-        V = self._V_matrix(phi)
-        U = np.identity(5)
-        U[:3,:3] = O
-        U[:3,3] = np.ravel(dt* J @ a)
-        U[:3,4] = np.ravel(dt**2 * V @ a)
-        U[3,4] = dt
-        return U
-
-    def _U_matrix_inv(self, u:Imu, dt: float):
-        phi = u.gyro * dt
-        O = SO3.Exp(phi)
-        V = self._V_matrix(phi)
-        J = SO3.left_jacobian(phi)
-        a = u.accel.reshape((-1,1))
-        U_inv = np.identity(5)
-        U_inv[:3,:3] = O.T
-        U_inv[:3,3] = np.ravel(-dt* O.T @ J @ a)
-        U_inv[:3,4] = np.ravel(dt**2 * O.T @ (J - V) @ a)
-        U_inv[3,4] = -dt
-        return U_inv
-
-    def _G_matrix(self, dt):
-        G = np.identity(5)
-        G[:3, 3] = dt * self._g
-        G[:3, 4] = -0.5 * dt**2 * self._g
-        G[3, 4] = -dt
-        return G
-
-    def _G_matrix_inv(self, dt):
-        G_inv = np.identity(5)
-        G_inv[:3, 3] = -dt * self._g
-        G_inv[:3, 4] = - 0.5 * dt**2 * self._g
-        G_inv[3, 4] = dt
-        return G_inv
-
-    # @staticmethod
-    # def _adjoint_U(U):
-    #     O = U[:3,:3]
-    #     dt = U[3,4]
-    #     Ja = U[:3,3].reshape((-1,1))/dt
-    #     Va = U[:3,4].reshape((-1,1))/dt**2
-    #     Ad = np.zeros((9,9))
-    #     Ad[:3,:3] = O
-    #     Ad[3:6,:3] = dt*SO3.wedge(Ja) @ O
-    #     Ad[3:6,3:6] = O
-
-    #     Ad[6:9,:3] = -dt**2*SO3.wedge(Ja - Va) @ O
-    #     Ad[6:9,3:6] = -dt * O 
-    #     Ad[6:9,6:9] = O
-    #     return Ad
-
-    @staticmethod
-    def _adjoint_U(U):
-        O = U[:3,:3]
-        c = U[3,4]
-        a = U[:3,3].reshape((-1,1))
-        b = U[:3,4].reshape((-1,1))
-        Ad = np.zeros((9,9))
-        Ad[:3,:3] = O
-        Ad[3:6,:3] = SO3.wedge(a) @ O
-        Ad[3:6,3:6] = O
-
-        Ad[6:9,:3] = -SO3.wedge(c*a - b) @ O
-        Ad[6:9,3:6] = -c * O 
-        Ad[6:9,6:9] = O
-        return Ad
